+++ TableProMobile/TableProMobile/Models/ConnectedTab.swift
@@ -7,5 +7,5 @@ enum ConnectedTab: String, CaseIterable, Sendable {
     case tables
     case query
     case history
-    case settings
+    case info
 }